Prevent bittensor import in API
